add testing utils target
