+++ Tests/IntegrationTests/Helpers.swift
@@ -1,27 +1,6 @@
+
 import Foundation
 @testable import WalletConnect
-
-let defaultTimeout: TimeInterval = 5.0
-
-extension String {
-    static func randomTopic() -> String {
-        "\(UUID().uuidString)\(UUID().uuidString)".replacingOccurrences(of: "-", with: "").lowercased()
-    }
-}
-
-extension Result {
-    
-    var isSuccess: Bool {
-        if case .success = self { return true }
-        return false
-    }
-    
-    var isFailure: Bool {
-        if case .failure = self { return true }
-        return false
-    }
-}
-
 extension WCRequest {
     
     var isPairingApprove: Bool {
