+++ Tests/TestingUtils/Helpers.swift
@@ -1,0 +1,23 @@
+import Foundation
+
+public let defaultTimeout: TimeInterval = 5.0
+
+public extension String {
+    static func randomTopic() -> String {
+        "\(UUID().uuidString)\(UUID().uuidString)".replacingOccurrences(of: "-", with: "").lowercased()
+    }
+}
+
+public extension Result {
+    
+    var isSuccess: Bool {
+        if case .success = self { return true }
+        return false
+    }
+    
+    var isFailure: Bool {
+        if case .failure = self { return true }
+        return false
+    }
+}
+
